+++ audioelementplugin/src/AudioElementPluginProcessor.cpp
@@ -27,25 +27,9 @@
 #include "processors/mix_monitoring/TrackMonitorProcessor.h"
 #include "processors/panner/Panner3DProcessor.h"
 #include "processors/routing/RoutingProcessor.h"
+#include "substream_rdr/substream_rdr_utils/Speakers.h"
 
 int AudioElementPluginProcessor::instanceId_ = 0;
-
-// Helpers for host layout negotiation (anonymous namespace for internal
-// linkage)
-namespace {
-inline bool isNamedBed(const juce::AudioChannelSet& s) {
-  return s == juce::AudioChannelSet::stereo() ||
-         s == juce::AudioChannelSet::create5point1() ||
-         s == juce::AudioChannelSet::create7point1() ||
-         s == juce::AudioChannelSet::create7point1point2() ||
-         s == juce::AudioChannelSet::create7point1point4();
-}
-inline bool isSymmetricDiscrete(const juce::AudioChannelSet& s) {
-  if (!s.isDiscreteLayout()) return false;
-  const int n = s.size();
-  return n >= 1 && n <= 16;  // covers 1,2,6,8,10,12 etc.
-}
-}  // namespace
 
 AudioElementPluginProcessor::AudioElementPluginProcessor()
     // For AU builds: use host-wide layout only for Logic Pro, not Premiere Pro
@@ -116,7 +100,7 @@ bool AudioElementPluginProcessor::isBusesLayoutSupported(
     const auto in = layouts.getMainInputChannelSet();
     const auto out = layouts.getMainOutputChannelSet();
     if (in.isDisabled() || out.isDisabled()) return false;
-    return isNamedBed(in) || isSymmetricDiscrete(in);
+    return Speakers::isNamedBed(in) || Speakers::isSymmetricDiscrete(in);
   }
 
   // prevent REAPER from downsizing the output channel set when
